Correct architectures value in library.properties
The previous architectures value caused the Arduino IDE to display a warning when the library is compiled:

WARNING: library Arduino_ConnectionHandler claims to run on (SAMD, ESP12, ESP32, esp8266) architecture(s) and may be incompatible with your current board which runs on (samd) architecture(s).

The previous architectures value caused the library's examples to be placed under the File > Examples > INCOMPATIBLE menu.