Run compile-examples CI action for ESP8266 board
The ESP8266 core for Arduino has Python 3 as a dependency. On Linux, instead of installing their own copy of Python 3, they assume it is already installed at /usr/bin/python3 and only install a symlink to it. The Docker container used by the compile-examples action does not have Python 3 installed, so this required using a wrapper script as an alternate entrypoint when compiling for an ESP8266 board. The wrapper script installs Python 3 and then passes the arguments on to the standard entrypoint script.