Use default branch of ArduinoCore-API in CI workflow
Previously, the `namespace_arduino` branch of `arduino/ArduinoCore-API` was used in the core library of the development
version of the Mbed OS Boards platform used in the "Compile Examples" CI workflow. The default branch of
`arduino/ArduinoCore-API` is now the latest and greatest version, so the `namespace_arduino` branch should no longer be
used for the Mbed OS Boards platform's core library.