Update badges markup in readme
Since the time the original badges were added, GitHub added support for referencing the workflows via file name in
addition to the previous approach of using the `name` value. This links to the specific workflow's runs, rather than a
search for the name which might return multiple results. It is also more lightweight in terms of syntax.