Use report size deltas action from dedicated repository
The arduino/actions repository is used for hosting GitHub Actions that are in an experimental state. The report size
deltas action has since been deemed valuable and mature enough to be moved to a dedicated repository. The action at the
previous location is preserved for now for the sake of backwards compatibility, but is deprecated and unmaintained. All
workflows should be updated to using the arduino/report-size-deltas action.