fix: Notecard_ConnectionHandler Power Profile
There is a bug in code that specifically targets the Wireless for Arduino Opta expansion. The battery backup on the device is a Lithium Ion Capacitor (LIC), but the Lithium Polymer (LiPo) profile was provided.

The Notecard will work correctly in this state, but it the LIC is designed to support lower voltages than a LiPo. The improvement provided here will allow the Notecard to remain powered for a longer duration.